fix: address server-side code review findings in region-split
- model.ts: add 15s (configurable) fetch timeout via AbortController;
  timeout raises a readable "model request timed out after Nms" error
  and is treated as a network-layer failure (no retry), matching the
  existing non-2xx behavior.
- server.ts: /api/model-config/test no longer falls back to the saved
  plaintext API key when the request's baseUrl differs from the saved
  baseUrl, closing a key-exfiltration path.
- types.ts: rectSchema now requires integer x/y/w/h so fractional
  bounds (which can pass checkInvariants but crash sharp.extract) are
  rejected at write time, matching the integer check already used by
  GET /image?rect=.
- analyze.ts: analyzeProject no longer wipes candidateLines to [] when
  detectLines is omitted; it now reuses doc.candidateLines computed at
  upload time, preserving split-snap behavior and avoiding a redundant
  full-image rowStats pass.
- analyze.ts/server.ts: non-image uploads now return 400 with a
  readable message (InvalidImageError) instead of a bare 500, and the
  project directory is only created after the image is validated so
  failed uploads don't leave an empty directory behind.
- analyze.ts: disk-read failures while loading images for the model
  (ENOENT, "Input file is missing: ...") are converted to a generic,
  path-free error before they can reach a 502 response body; model-layer
  error messages (timeout, HTTP status, parse failure) pass through
  unchanged.

All changes are TDD: each fix has a failing test added first. 90/90
tests pass in packages/region-split, 148/148 across the ui workspace,
typecheck clean. Scope limited to packages/region-split; apps/region-split-ui
untouched.

+++ ui/packages/region-split/src/types.ts
@@ -11,7 +11,7 @@ export type RegionType = (typeof regionTypes)[number];
 export interface Rect { x: number; y: number; w: number; h: number }
 
 export const rectSchema = z.object({
-  x: z.number(), y: z.number(), w: z.number().positive(), h: z.number().positive(),
+  x: z.number().int(), y: z.number().int(), w: z.number().int().positive(), h: z.number().int().positive(),
 });
 
 export const regionSchema = z.object({
